requirements.txt
python-jobspy==1.1.80
